Split type widening tests in multiple suites
